Fix multiple definition

+++ 2.0/include/plink2_bits.h
@@ -74,7 +74,7 @@ uintptr_t AdvTo0Bit(const uintptr_t* bitarr, uintptr_t loc);
 
 uint32_t AdvBoundedTo1Bit(const uintptr_t* bitarr, uint32_t loc, uint32_t ceil);
 
-uint32_t AdvBoundedTo0Bit(const uintptr_t* bitarr, uint32_t loc, uint32_t ceil);
+uintptr_t AdvBoundedTo0Bit(const uintptr_t* bitarr, uintptr_t loc, uintptr_t ceil);
 
 uint32_t FindLast1BitBefore(const uintptr_t* bitarr, uint32_t loc);
